refactor(ui-test): apply all picker-uitest review feedback
Critical:
- SettingsScreen now routes through MoolahApp.toolbarButton(label:) with
  a typed UITestIdentifiers.Settings.cryptoTabTitle constant (was raw
  'Crypto' string). The lookup stays label-based because SwiftUI's Tab
  on macOS does not propagate .accessibilityIdentifier(_:) to the
  generated toolbar button — verified by inspection of the failure
  artefact tree, where the Tab buttons surface with no identifier. The
  new resolver method preserves the single-resolver invariant
  (UI_TEST_GUIDE §3 #5) by routing label-based lookups through MoolahApp
  alongside element(for:).

Important:
- AddTokenScreen.search waits for the typed value to propagate into the
  search field's accessibility value before returning so the debounce
  fires against the intended query.
- AddTokenScreen.selectResult waits for the row to stop being hittable
  as a minimal post-condition that the click registered (dismissal is
  still owned by waitForDismiss).
- Split UITestSeedCryptoOverrides into per-type files (CODE_GUIDE §2):
  PreloadedCryptoCatalog and PreloadedTokenResolutionClient now each
  live in their own file, leaving UITestSeedCryptoOverrides as just
  the seed dispatcher.
- SettingsView macOS HSplitView/TabView layout extracted to
  SettingsView+macOS.swift to keep the main type body under SwiftLint's
  type_body_length threshold after the new tab annotation. Sidebar
  toolbar HStack closure factored into addProfileButton /
  removeProfileButton / importProfileButton helpers so the closure
  body stays under closure_body_length.

Minor:
- cryptoCatalogPreloaded seed comment cross-references
  UITestFixtures.CryptoCatalogPreloaded for the complete fixture so
  seed.txt readers don't need a cross-file lookup.
- Confirmed InstrumentPicker.sheet and InstrumentPicker.searchField
  identifiers are present on the production view
  (Shared/Views/InstrumentPickerSheetCore.swift lines 28, 80, 112).
- Added // MARK: - section headers to ProfileSession+Factories.swift
  (Market Data Services / Backend / Registry Wiring / Domain Stores /
  Import Pipeline / Folder Watch Services).
