+++ App/ProfileSession+Factories.swift
@@ -6,6 +6,8 @@ import OSLog
 import SwiftData
 
 extension ProfileSession {
+  // MARK: - Market Data Services
+
   /// Bundle of the external market-data services a profile session depends
   /// on: fiat exchange rates, stock prices, and crypto prices. Returned
   /// from `makeMarketDataServices` so `init` can assign each field in one
@@ -68,6 +70,8 @@ extension ProfileSession {
     )
   }
 
+  // MARK: - Backend
+
   /// Builds the `BackendProvider` for the profile based on its backend type.
   /// iCloud profiles get the full conversion service (stock + crypto + fiat);
   /// remote profiles use their own internal fiat-only conversion.
@@ -115,6 +119,8 @@ extension ProfileSession {
           cryptoPrices: cryptoPrices))
     }
   }
+
+  // MARK: - Registry Wiring
 
   /// Bundle of the optional instrument-registry pieces: only CloudKit
   /// profiles expose a registry, crypto token store, search service,
@@ -222,6 +228,8 @@ extension ProfileSession {
     }
   }
 
+  // MARK: - Domain Stores
+
   /// Bundle of the per-profile domain stores. Returned from
   /// `makeDomainStores` so `ProfileSession.init` can assign each stored
   /// property in one step without inlining every constructor call.
@@ -277,6 +285,8 @@ extension ProfileSession {
     )
   }
 
+  // MARK: - Import Pipeline
+
   /// Bundle of the full CSV import pipeline: the `ImportStore`, the import-rule
   /// store, and the three folder-watch pieces. Returned from `makeImportPipeline`
   /// so `ProfileSession.init` can assign all five fields in one step.
@@ -320,6 +330,8 @@ extension ProfileSession {
       watcher: folderWatch.watcher
     )
   }
+
+  // MARK: - Folder Watch Services
 
   /// Bundle of the services that make up folder-watch ingestion for a
   /// profile: the on-disk `ImportPreferences`, the catch-up `FolderScanService`,
